MoM_Backend/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
filelock==3.11.0
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.13.4
hyperframe==5.2.0
idna==2.10
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.1
nltk==3.8.1
numpy==1.24.2
packaging==23.0
pathlib==1.0.1
PyJWT==1.7.1
pytz==2023.3
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
rfc3986==1.5.0
sentencepiece==0.1.98
sniffio==1.3.0
sqlparse==0.4.3
sympy==1.11.1
tokenizers==0.13.3
tomli==2.0.1
torch==2.0.0
tqdm==4.65.0
transformers==4.27.4
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
yapf==0.33.0
